Changed some column types

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -33,7 +33,7 @@ class CreateUsersTable extends Migration
 
             $table->integer('pin_number')->unsigned();
 
-            $table->integer('type')->unsigned();
+            $table->tinyInteger('type')->unsigned();
 
             $table->timestamps();
