+++ database/migrations/2022_04_18_092622_create_sale_return_transactions_table.php
@@ -36,7 +36,7 @@ class CreateSaleReturnTransactionsTable extends Migration
 
             $table->dateTime('finalized_at');
 
-            $table->integer('verification_status')->unsigned()->default(2);
+            $table->tinyInteger('verification_status')->unsigned()->default(2);
 
             $table->string('verification_note')->nullable();
